modifysauce with verification

+++ backend/controllers/sauces.js
@@ -25,27 +25,41 @@ exports.getOneSauce = (req, res, next) => {
 
 // Update
 exports.modifySauce = (req, res, next) => {
-  const sauceObject = req.file ?
-    {
-      ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body }
-  Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
-    .catch(error => res.status(400).json({ error }))
+	// Get parameters & data
+  Sauce.findOne({ _id: req.params.id })
+    .then(sauce => {
+      if (sauce.userId !== req.auth.userId) {
+        res.status(403).json({ message: "Vous n'êtes pas le proppriétaire de cette sauce"})
+      } else {
+        // If file 
+        const sauceObject = req.file ? {
+          ...JSON.parse(req.body.sauce),
+          imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        } : { ...req.body } // If no file
 
+        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+        .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
+        .catch(error => res.status(400).json({ error }))
+      }
+    })
+    .catch(error => res.status(401).json({ error }))
 }
 
 // Delete
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
-      const filename = sauce.imageUrl.split('/images/')[1]
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Sauce supprimée !'}))
-          .catch(error => res.status(400).json({ error }))
-      })
+    	if(sauce.userId !== req.auth.userId) {
+    		res.status(403).json({ message: "Vous n'êtes pas le propriétaire de la sauce" })
+    	}
+    	else {
+	      const filename = sauce.imageUrl.split('/images/')[1]
+	      fs.unlink(`images/${filename}`, () => {
+	        Sauce.deleteOne({ _id: req.params.id })
+	          .then(() => res.status(200).json({ message: 'Sauce supprimée !'}))
+	          .catch(error => res.status(400).json({ error }))
+	      })
+    	}
     })
     .catch(error => res.status(500).json({ error }))
 }
